Putting functions into files instead of main.cpp
