+++ VentilatorControl/include/includes.h
@@ -2,16 +2,20 @@
 #define __VENTILATOR_INCLUDES_H__
 
 #include "Arduino.h"
+
 #include "Wire.h"
 #include "TCA9548A.h"
 #include "AllSensors_DLHR.h"
 #include "Servo.h"
 #include "PID_v1.h"
 #include "Adafruit_NeoPixel.h"
+
 #include "Types.h"
 #include "variables.h"
-#include "helpers.h"
 #include "Pins.h"
 #include "Constants.h"
+#include "timing.h"
+#include "compute_comms.h"
+#include "helpers.h"
 
 #endif  // __VENTILATOR_INCLUDES_H__
