+++ VentilatorControl/src/compute_comms.h
@@ -1,0 +1,8 @@
+// Compute (currently Raspberry Pi Model B+) Bidirectional Communication Functions
+
+#ifndef __COMPUTE_COMMS_H__
+#define __COMPUTE_COMMS_H__
+
+void get_values_from_raspberry_pi (void);
+
+#endif // __COMPUTE_COMMS_H__
