+++ VentilatorControl/src/timing.cpp
@@ -1,0 +1,12 @@
+#include "includes.h"
+
+void breath_cycle_timer_reset (bool hardreset)
+// void breath_cycle_timer_reset (bool hardreset = false)
+{
+  CurrTimeInCycleMilliseconds = 0;
+  CycleStartTimeFromSysClockMilliseconds = millis();
+  if(hardreset)
+  {
+    ControlLoopStartTimeMilliseconds = CycleStartTimeFromSysClockMilliseconds;
+  }
+}
