+++ VentilatorControl/src/timing.h
@@ -1,0 +1,6 @@
+#ifndef __TIMING_H__
+#define __TIMING_H__
+
+void breath_cycle_timer_reset (bool hardreset = false);
+
+#endif // __TIMING_H__
